requirements.txt
anthropic>=0.30.0
pyyaml>=6.0
networkx>=3.0
tree-sitter>=0.20.0
tree-sitter-python>=0.20.0
tree-sitter-javascript>=0.20.0
tree-sitter-java>=0.20.0
rich>=13.0.0
